+++ strelyaev.roman/T3/polygon.hpp
@@ -1,0 +1,29 @@
+#ifndef POLYGON_HPP
+#define POLYGON_HPP
+#include <iosfwd>
+#include <vector>
+
+namespace strelyaev
+{
+  struct Point
+  {
+    int x;
+    int y;
+  };
+  std::istream& operator>>(std::istream&, Point&);
+
+  struct Polygon
+  {
+    std::vector< Point > points;
+  };
+  std::istream& operator>>(std::istream&, Polygon&);
+  size_t size_getter(const Polygon&);
+  int get_x(const Point&);
+  int get_y(const Point&);
+  double get_area(const Polygon&);
+  bool operator==(const Polygon&, const Polygon&);
+  bool operator==(const Point&, const Point&);
+}
+
+#endif
+
